Unit Test hardening - Teardown Code now tests for file existence before unlinking.

+++ test/soap/marshal/test_digraph.rb
@@ -47,7 +47,7 @@ class TestDigraph < Test::Unit::TestCase
 
   def teardown
     if File.exist?("digraph_marshalled_string.soap")
-      File.unlink("digraph_marshalled_string.soap")
+      File.unlink("digraph_marshalled_string.soap") if File.file?("digraph_marshalled_string.soap")
     end
   end
 end
