+++ test/wsdl/document/array/test_array.rb
@@ -65,9 +65,9 @@ class TestArray < Test::Unit::TestCase
   def teardown
     teardown_server if @server
     unless $DEBUG
-      File.unlink(pathname('double.rb'))
-      File.unlink(pathname('doubleMappingRegistry.rb'))
-      File.unlink(pathname('doubleDriver.rb'))
+      File.unlink(pathname('double.rb')) if File.file?(pathname('double.rb'))
+      File.unlink(pathname('doubleMappingRegistry.rb')) if File.file?(pathname('doubleMappingRegistry.rb'))
+      File.unlink(pathname('doubleDriver.rb')) if File.file?(pathname('doubleDriver.rb'))
     end
     @client.reset_stream if @client
   end
